+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Fixes
 
 - Fix misbehaving stdio connector when stdio is a pipe
-- fix Docker entrypoint script argument ordering
+- Fix Docker entrypoint script argument ordering
+- Fix id encoding for discord connector
 
 ## [0.12.0]