+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.13 UNRELEASED 
+
+### Breaking Changes
+
+- update grok to 0.2, optional patterns are now omitted no longer ""
+
 ## Unreleased
 ### New features